moved command line parser to own module
